cleanup - upgrading dependencies - seperating components
