+++ package.json
@@ -10,9 +10,9 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "0xsequence": "^1.9.36",
-    "@0xsequence/kit": "^2.8.11",
-    "@0xsequence/kit-connectors": "^2.8.11",
+    "0xsequence": "1.9.36",
+    "@0xsequence/kit": "2.8.11",
+    "@0xsequence/kit-connectors": "2.8.11",
     "@tanstack/react-query": "^5.50.1",
     "ethers": "5.7.2",
     "react": "^18.3.1",
@@ -23,12 +23,12 @@
   "devDependencies": {
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^7.13.1",
-    "@typescript-eslint/parser": "^7.13.1",
+    "@typescript-eslint/eslint-plugin": "^7.17.0",
+    "@typescript-eslint/parser": "^7.17.0",
     "@vitejs/plugin-react": "^4.3.1",
     "eslint": "^8.57.0",
     "eslint-plugin-react-hooks": "^4.6.2",
-    "eslint-plugin-react-refresh": "^0.4.7",
+    "eslint-plugin-react-refresh": "^0.4.9",
     "prettier": "^3.3.3",
     "typescript": "^5.5.4",
     "vite": "^5.3.4"
