+++ src/views/components/NotConnected.tsx
@@ -1,0 +1,20 @@
+import { useOpenConnectModal } from "@0xsequence/kit";
+
+const NotConnected = () => {
+  const { setOpenConnectModal } = useOpenConnectModal();
+
+  const onClickConnect = () => {
+    setOpenConnectModal(true);
+  };
+
+  return (
+    <>
+      <p>Not connected</p>
+      <div className="card">
+        <button onClick={onClickConnect}>Connect</button>
+      </div>
+    </>
+  );
+};
+
+export default NotConnected;
